Break Nav into Presentational and Container components
